Remove stuff about ejecting
Ejecting create-react-app is likely to be a confusing experience
for a React newb and isn't actually necessary for doing this kata.
(Furthermore there were two contradictory pieces of instruction
about ejecting, one with `npm` and the other with `yarn`.)

Per the documentation, Create React App already has a
standard location for a test setup file, `src/setupTests.js`, and
even describes putting this same Enzyme setup in it here:
https://facebook.github.io/create-react-app/docs/running-tests#option-1-shallow-rendering

This simplifies the kata a little bit.